CRM：完善【排行版】功能

+++ yudao-module-crm/yudao-module-crm-biz/src/main/java/cn/iocoder/yudao/module/crm/controller/admin/statistics/vo/rank/CrmStatisticsRankRespVO.java
@@ -3,6 +3,8 @@ package cn.iocoder.yudao.module.crm.controller.admin.statistics.vo.rank;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 
 @Schema(description = "管理后台 - CRM 排行榜统计 Response VO")
 @Data
@@ -17,14 +19,15 @@ public class CrmStatisticsRankRespVO {
     @Schema(description = "部门名称", requiredMode = Schema.RequiredMode.REQUIRED, example = "1")
     private String deptName;
 
-    // TODO @芋艿：需要改下，金额是 bigdecimal
     /**
      * 数量是个特别“抽象”的概念，在不同排行下，代表不同含义
-     * <p>
+     *
      * 1. 金额：合同金额排行、回款金额排行
      * 2. 个数：签约合同排行、产品销量排行、产品销量排行、新增客户数排行、新增联系人排行、跟进次数排行、跟进客户数排行
+     *
+     * 为什么使用 BigDecimal 的原因：
      */
     @Schema(description = "数量", requiredMode = Schema.RequiredMode.REQUIRED, example = "1")
-    private Integer count;
+    private BigDecimal count;
 
 }
